[#183] move changed-id store into event store
