+++ src/shared/constants/pages.ts
@@ -1,4 +1,4 @@
-import {EVENT_TYPES} from "../types";
+import {EVENT_TYPES} from "../types/events";
 
 export const PAGE_TYPES = {
   ...EVENT_TYPES,
