+++ src/shared/stores/events/index.ts
@@ -1,0 +1,2 @@
+export {useEventStore} from './events-store'
+export * from './types'
