+++ src/shared/stores/events/types.ts
@@ -1,0 +1,4 @@
+import type {EventId} from "../../types/events";
+import type { TEventsGroup } from "../../types/pages";
+
+export type TEventsCachedIdsMap = Record<TEventsGroup, EventId[]>;
