+++ src/shared/stores/index.ts
@@ -1,4 +1,3 @@
-export * from './cached-ids'
 export * from './connections'
 export * from './events'
 export * from './locked-ids'
