+++ src/shared/types/index.ts
@@ -3,3 +3,4 @@ export * from "./profile";
 export * from "./generics";
 export * from "./local-storage";
 export * from "./partials";
+export * from "./pages";
