+++ src/shared/types/pages.ts
@@ -1,0 +1,4 @@
+import { PAGE_TYPES } from "../constants/pages";
+import type {OneOfValues} from "./generics";
+
+export type TEventsGroup = OneOfValues<typeof PAGE_TYPES>
